Refactored file paths
